th7

+++ declaration.go
@@ -4,7 +4,7 @@ type OldRepInfo struct {
 	Date    string
 	Player  string
 	Char    string
-	Score   uint32
+	Score   int64
 	Rank    string
 	Version string
 	Drop    float32 // 处理落率（百分比）
